refactor: 예외 처리 로직 분리
